refactor: 레거시 코드 삭제 및 이름 변경
