+++ server/src/main/java/com/dialog/server/dto/response/DiscussionDetailResponse.java
@@ -4,68 +4,100 @@ import com.dialog.server.domain.Category;
 import com.dialog.server.domain.Discussion;
 import com.dialog.server.domain.DiscussionParticipant;
 import com.dialog.server.domain.OfflineDiscussion;
+import com.dialog.server.domain.OnlineDiscussion;
 import com.dialog.server.domain.ProfileImage;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public record DiscussionDetailResponse(
         Long id,
-        String title,
-        String content,
-        LocalDateTime startAt,
-        LocalDateTime endAt,
-        String place,
-        Category track,
-        int participantCount,
-        int maxParticipantCount,
-        String summary,
-        LocalDateTime createdAt,
-        LocalDateTime modifiedAt,
-        long likeCount,
-        boolean isBookmarked,
-        AuthorResponse author,
-        List<ParticipantResponse> participants
+        DiscussionType discussionType,
+        CommonDiscussionInfo commonDiscussionInfo,
+        OfflineDiscussionInfo offlineDiscussionInfo,
+        OnlineDiscussionInfo onlineDiscussionInfo
 ) {
-    public static DiscussionDetailResponse of(Discussion discussion,
-                                              long likeCount,
-                                              List<DiscussionParticipant> participants,
-                                              ProfileImage profileImage) {
-        if (discussion instanceof OfflineDiscussion offlineDiscussion) {
-            return new DiscussionDetailResponse(
-                    offlineDiscussion.getId(),
-                    offlineDiscussion.getTitle(),
-                    offlineDiscussion.getContent(),
-                    offlineDiscussion.getStartAt(),
-                    offlineDiscussion.getEndAt(),
-                    offlineDiscussion.getPlace(),
-                    offlineDiscussion.getCategory(),
-                    offlineDiscussion.getParticipantCount(),
-                    offlineDiscussion.getMaxParticipantCount(),
-                    offlineDiscussion.getSummary(),
-                    offlineDiscussion.getCreatedAt(),
-                    offlineDiscussion.getModifiedAt(),
-                    likeCount,
-                    false,
-                    toAuthorResponse(offlineDiscussion, profileImage),
-                    toParticipantResponse(participants)
-            );
-        }
-        throw new IllegalArgumentException("Unsupported discussion type");
-    }
 
-    private static AuthorResponse toAuthorResponse(Discussion discussion, ProfileImage profileImage) {
-        return new AuthorResponse(
-                discussion.getAuthor().getId(),
-                discussion.getAuthor().getNickname(),
-                profileImage == null ? null : ProfileImageResponse.from(profileImage)
+    public static DiscussionDetailResponse fromOfflineDiscussion(
+            OfflineDiscussion offlineDiscussion,
+            long likeCount,
+            ProfileImage profileImage
+    ) {
+
+        CommonDiscussionInfo commonInfo = CommonDiscussionInfo.from(
+                offlineDiscussion,
+                profileImage,
+                likeCount
+        );
+
+        return new DiscussionDetailResponse(
+                offlineDiscussion.getId(),
+                DiscussionType.OFFLINE,
+                commonInfo,
+                new OfflineDiscussionInfo(
+                        offlineDiscussion.getStartAt(),
+                        offlineDiscussion.getEndAt(),
+                        offlineDiscussion.getPlace(),
+                        offlineDiscussion.getParticipantCount(),
+                        offlineDiscussion.getMaxParticipantCount(),
+                        ParticipantResponse.toParticipantResponse(offlineDiscussion.getDiscussionParticipants())
+                ),
+                null
         );
     }
 
-    private static List<ParticipantResponse> toParticipantResponse(List<DiscussionParticipant> participants) {
-        return participants.stream().map(participant -> new ParticipantResponse(
-                participant.getParticipant().getId(),
-                participant.getParticipant().getNickname()
-        )).toList();
+    public static DiscussionDetailResponse fromOnlineDiscussion(
+            OnlineDiscussion onlineDiscussion,
+            long likeCount,
+            ProfileImage profileImage
+    ) {
+
+        CommonDiscussionInfo commonInfo = CommonDiscussionInfo.from(
+                onlineDiscussion,
+                profileImage,
+                likeCount
+        );
+
+        return new DiscussionDetailResponse(
+                onlineDiscussion.getId(),
+                DiscussionType.ONLINE,
+                commonInfo,
+                null,
+                new OnlineDiscussionInfo(onlineDiscussion.getEndDate())
+        );
+    }
+
+    public enum DiscussionType {
+        ONLINE, OFFLINE
+    }
+
+    public record CommonDiscussionInfo(
+            String title,
+            String content,
+            String summary,
+            Category category,
+            LocalDateTime createdAt,
+            LocalDateTime modifiedAt,
+            long likeCount,
+            AuthorResponse author
+    ) {
+        private static CommonDiscussionInfo from(
+                Discussion discussion,
+                ProfileImage profileImage,
+                long likeCount
+        ) {
+            return new CommonDiscussionInfo(
+                    discussion.getTitle(),
+                    discussion.getContent(),
+                    discussion.getSummary(),
+                    discussion.getCategory(),
+                    discussion.getCreatedAt(),
+                    discussion.getModifiedAt(),
+                    likeCount,
+                    AuthorResponse.toAuthorResponse(discussion, profileImage)
+            );
+        }
     }
 
     public record AuthorResponse(
@@ -73,12 +105,13 @@ public record DiscussionDetailResponse(
             String name,
             ProfileImageResponse profileImage
     ) {
-    }
-
-    public record ParticipantResponse(
-            Long id,
-            String name
-    ) {
+        private static AuthorResponse toAuthorResponse(Discussion discussion, ProfileImage profileImage) {
+            return new AuthorResponse(
+                    discussion.getAuthor().getId(),
+                    discussion.getAuthor().getNickname(),
+                    profileImage == null ? null : ProfileImageResponse.from(profileImage)
+            );
+        }
     }
 
     public record ProfileImageResponse(
@@ -91,5 +124,37 @@ public record DiscussionDetailResponse(
                     profileImage.getCustomImageUri()
             );
         }
+    }
+
+    public record OfflineDiscussionInfo(
+            @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+            LocalDateTime startAt,
+            @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+            LocalDateTime endAt,
+            String place,
+            int participantCount,
+            int maxParticipantCount,
+            List<ParticipantResponse> participants
+    ) {
+    }
+
+    public record ParticipantResponse(
+            Long id,
+            String name
+    ) {
+        private static List<ParticipantResponse> toParticipantResponse(List<DiscussionParticipant> participants) {
+            return participants.stream()
+                    .map(participant -> new ParticipantResponse(
+                            participant.getParticipant().getId(),
+                            participant.getParticipant().getNickname()
+                    ))
+                    .toList();
+        }
+    }
+
+    public record OnlineDiscussionInfo(
+            @JsonFormat(pattern = "yyyy-MM-dd")
+            LocalDate endDate
+    ) {
     }
 }
